rework application pre-bootstrap setup
- extract keycloak setup
- setup branding and keycloak simultaneously
- handle product.json loading error
