+++ src/app/index.styl
@@ -269,7 +269,7 @@ md-tabs-wrapper
   font-size 12px
   border-radius 2px
 
-.keycloak-error-backdrop
+.prebootstrap-error-backdrop
   background-color $black-color
   position absolute
   top 0
@@ -284,7 +284,7 @@ md-tabs-wrapper
     width 100px
     margin-left -50px
 
-.keycloak-error
+.prebootstrap-error-container
   background-color $white-color
   color $error-color
   padding 50px 0
