+++ src/assets/branding/product.json
@@ -1,36 +1,1 @@
-{
-  "title": "Eclipse Che ",
-  "name": "Eclipse Che",
-  "logoFile": "che-logo.svg",
-  "logoTextFile": "che-logo-text.svg",
-  "favicon": "favicon.ico",
-  "loader": "loader.svg",
-  "websocketContext": "/api/websocket",
-  "helpPath": "https://www.eclipse.org/che/",
-  "helpTitle": "Community",
-  "supportEmail": "che-dev@eclipse.org",
-  "oauthDocs": "Configure OAuth in the che.properties file.",
-  "workspace": {
-    "priorityStacks": [
-      "Java",
-      "Java-MySQL",
-      "Blank"
-    ],
-    "defaultStack": "java-mysql"
-  },
-  "cli": {
-    "configName": "che.env",
-    "name": "CHE"
-  },
-  "docs": {
-    "devfile": "https://www.eclipse.org/che/docs/che-7/making-a-workspace-portable-using-a-devfile/",
-    "workspace": "https://www.eclipse.org/che/docs/che-7/workspaces-overview/",
-    "factory": "https://www.eclipse.org/che/docs/factories-getting-started.html",
-    "organization": "https://www.eclipse.org/che/docs/organizations.html",
-    "converting": "https://www.eclipse.org/che/docs/che-7/converting-a-che-6-workspace-to-a-che-7-devfile/",
-    "certificate": "https://www.eclipse.org/che/docs/che-7/importing-certificates-to-browsers/",
-    "general": "https://www.eclipse.org/che/docs/che-7",
-    "storageTypes": "https://www.eclipse.org/che/docs/che-7/configuring-storage-types/",
-    "webSocketTroubleshooting": "https://www.eclipse.org/che/docs/che-7/troubleshooting-network-problems/#troubleshooting-websocket-secure-connections_troubleshooting-network-problems"
-  }
-}
+{}
